I1 - Fix All warnings. Project compiles succesfully.

+++ FreeRTOSV8.2.3/FreeRTOS/Demo/WIN32-MingW/main.c
@@ -191,7 +191,7 @@ void vAssertCalled(unsigned long ulLine, const char * const pcFileName)
 	(void)ulLine;
 	(void)pcFileName;
 
-	printf("ASSERT! Line %d, file %s\r\n", ulLine, pcFileName);
+	printf("ASSERT! Line %d, file %s\r\n", (int)ulLine, pcFileName);
 
 	taskENTER_CRITICAL();
 	{
